+++ packages/custom_offer_carousel/lib/src/colors.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class CustomColors {
+  CustomColors._();
+
+  static const Color navyBlue2 = Color(0xFF15376A);
+  static const Color lightGrey = Color(0xFFDBDBDB);
+  static const Color lightestBlue = Color(0xFFD2DDF1);
+}