+++ packages/custom_offer_carousel/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Offer's Carousel Options
+- Indicator Options